# Add randomized C++ solution