--- a/c++-shuffle/abcdefghppp.cpp
+++ b/c++-shuffle/abcdefghppp.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 
 #define X(a, b) (x[a]*10+x[b])
-int x[] = { 0, 1, 2, 3, 4, 5, 6, 7, 8, 9 };
+int x[] = { 8, 5, 4, 6, 3, 9, 7, 2, 1, 0 }; // I know this will be fast :)
 
 int correct() {
     int ab = X(0, 1), cd = X(2, 3), ef = X(4, 5), gh = X(6, 7), ppp = x[8] * 111;