--- a/c++-shuffle/abcdefghppp.cpp
+++ b/c++-shuffle/abcdefghppp.cpp
@@ -1,10 +1,14 @@
 #include <cstdio>
+#include <cstdlib>
+#include <ctime>
 #include <algorithm>
+
+#define MAX_TRIALS 5516431 // 99.95% getting correct answer
 
 using namespace std;
 
 #define X(a, b) (x[a]*10+x[b])
-int x[] = { 8, 5, 4, 6, 3, 9, 7, 2, 1, 0 }; // I know this will be fast :)
+int x[] = { 0, 1, 2, 3, 4, 5, 6, 7, 8, 9 };
 
 int correct() {
     int ab = X(0, 1), cd = X(2, 3), ef = X(4, 5), gh = X(6, 7), ppp = x[8] * 111;
@@ -16,12 +20,18 @@ int correct() {
 }
 
 int main() {
-    while (!correct()) {
+    int t = 1;
+    for (; t <= MAX_TRIALS && !correct(); t++) {
         random_shuffle(x, x + 10);
-    };
+    }
 
-    int ab = X(0, 1), cd = X(2, 3), ef = X(4, 5), gh = X(6, 7), ppp = x[8] * 111;
-    printf("%d - %d = %d, %d + %d = %d\n", ab, cd, ef, ef, gh, ppp);
+    if (correct()) {
+        printf("Answer found at #%d trial%s.\n", t, t == 1 ? "": "s");
+        int ab = X(0, 1), cd = X(2, 3), ef = X(4, 5), gh = X(6, 7), ppp = x[8] * 111;
+        printf("%d - %d = %d, %d + %d = %d\n", ab, cd, ef, ef, gh, ppp);
+    } else {
+        puts("I'm so stupid. :(");
+    }
 
     return 0;
 }