Make use of randomized algorithm learnt from the algorithm course.
It returns the first solution quite fast.

Before optimization:

```
$ time ./abcdefghppp
85 - 46 = 39, 39 + 72 = 111
./abcdefghppp  0.22s user 0.00s system 99% cpu 0.221 total
```
#### Edited

After optimization:

```
$ time ./abcdefghppp
85 - 46 = 39, 39 + 72 = 111
./abcdefghppp  0.00s user 0.00s system 64% cpu 0.003 total
```

The performance improved A LOT!