@gaplo917 2016-03-26
But my scala optimized solution only take 2ms (JMH Benchmark) for all solution :laughing: 

p.s  Macbook pro retina 15" 2012