@cherrry 2016-03-26
ok, let me optimize my code