@cherrry 2016-03-26
My optimized algorithm takes no time now, please kindly review again :joy: